Fixed comments and change int to float.

+++ FIXED_BATCH_3/minimum_number.py
@@ -1,20 +1,20 @@
-#Create a list
+# Create a list
 numbers_list = []
 
-#Use loop and try-except to allow users to input number and stop if the input is invalid
+# loop and try-except to allow users to input number and stop if the input is invalid
 while True:
     try:
-        number = int(input("Enter the number: "))
+        number = float(input("Enter the number: "))
         
-        #Update the list
+        # Update the list
         numbers_list.append(number)
 
     except ValueError:
         print("Invalid")
         break
 
-#Check which is the smallest number
+# Check which is the smallest number
 min_number = min(numbers_list)
         
-#Print the result
+# Print the result
 print(f"Result: {min_number}")
